+++ screens/LoginScreen.tsx
@@ -1,42 +1,30 @@
 import { Text, Button, Image, View, StyleSheet, TextInput } from "react-native";
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { SafeAreaView } from "react-native";
-import apiCalls from "../apiCalls/apiCalls";
-import { User, Attributes } from '../interface';
-import Tabs from '../navigation/tabs';
-import {COLORS} from '../constants/Colors';
 
-export default function LoginScreen() {
-	const [userName, setUserName] = useState('')
-	const [password, setPassword] = useState('')
-  const [user, setUser] = useState<User | null>(null);
-
-	const login = () => {
-		if(userName === 'RossInAVan' && password === 'Iamavan') {
-			apiCalls.getUser().then((data) => setUser(data.data));
-		}
-	}
+export default function LoginScreen({ login, setPassword, setUserName, userName, password } : {setPassword: any, setUserName: any, userName: string, password: string, login: any}) {
 
   return (
     <SafeAreaView style={styles.container} > 
-			{!user && <View>
+			<View>
 				<Text>Pom Pets</Text>
 				<Image style={styles.image} source={require('../assets/Pets/TomatoPet-Sad.png')} />
 				<TextInput  
+					autoCapitalize={"none"}
 					style={styles.input}
 					onChangeText={setUserName}
 					value={userName}
 					placeholder={'username'}
 				/>
 				<TextInput 
+					autoCapitalize={"none"}
 					style={styles.input}
 					onChangeText={setPassword}
 					value={password}
 					placeholder={'password'}
 				/>
 				<Button title="please click me to login" onPress={login} />
-			</View>}
-			{user && <Tabs />}
+			</View>
     </SafeAreaView> 
   );
 }
